be able to delete all of the events in a recurrance

+++ src/classes/ViewMeetings.php
@@ -10,13 +10,13 @@ class ViewMeetings {
         echo "<h3>";
         if ($tableType == 'me') {
             echo "Future Reservations";
-            $query = "SELECT res._id AS _id, first_name, last_name, name, room_number, start_time, end_time, date, rect.description as description FROM reservation res JOIN room r ON res.conference_room_id = r._id JOIN location l ON r.location_id = l._id JOIN user u ON res.user_id = u._id JOIN time_slot t ON res.time_slot_id = t._id JOIN recurrence rec on res.recurrence_id = rec._id JOIN recurrence_type rect on rec.recurrence_type_id = rect._id WHERE date >= CURDATE() AND user_id = " . $userId . " ORDER BY date";
+            $query = "SELECT res._id AS _id, first_name, last_name, name, room_number, start_time, end_time, date, recurrance_id, rect.description as description FROM reservation res JOIN room r ON res.conference_room_id = r._id JOIN location l ON r.location_id = l._id JOIN user u ON res.user_id = u._id JOIN time_slot t ON res.time_slot_id = t._id JOIN recurrence rec on res.recurrence_id = rec._id JOIN recurrence_type rect on rec.recurrence_type_id = rect._id WHERE date >= CURDATE() AND user_id = " . $userId . " ORDER BY date";
         } else if ($tableType == 'users') {
             echo "Created User's Reservations";
-            $query = "SELECT res._id AS _id, first_name, last_name, name, room_number, start_time, end_time, date, rect.description as description FROM reservation res JOIN room r ON res.conference_room_id = r._id JOIN location l ON r.location_id = l._id JOIN user u ON res.user_id = u._id JOIN time_slot t ON res.time_slot_id = t._id JOIN recurrence rec on res.recurrence_id = rec._id JOIN recurrence_type rect on rec.recurrence_type_id = rect._id WHERE date >= CURDATE() AND created_by_id = " . $userId . " ORDER BY date";
+            $query = "SELECT res._id AS _id, first_name, last_name, name, room_number, start_time, end_time, date, recurrance_id, rect.description as description FROM reservation res JOIN room r ON res.conference_room_id = r._id JOIN location l ON r.location_id = l._id JOIN user u ON res.user_id = u._id JOIN time_slot t ON res.time_slot_id = t._id JOIN recurrence rec on res.recurrence_id = rec._id JOIN recurrence_type rect on rec.recurrence_type_id = rect._id WHERE date >= CURDATE() AND created_by_id = " . $userId . " ORDER BY date";
         } else { // all users
             echo "All User's Reservations";
-            $query = "SELECT res._id AS _id, first_name, last_name, name, room_number, start_time, end_time, date, rect.description as description FROM reservation res JOIN room r ON res.conference_room_id = r._id JOIN location l ON r.location_id = l._id JOIN user u ON res.user_id = u._id JOIN time_slot t ON res.time_slot_id = t._id JOIN recurrence rec on res.recurrence_id = rec._id JOIN recurrence_type rect on rec.recurrence_type_id = rect._id WHERE date >= CURDATE()" . " ORDER BY date";
+            $query = "SELECT res._id AS _id, first_name, last_name, name, room_number, start_time, end_time, date, recurrance_id, rect.description as description FROM reservation res JOIN room r ON res.conference_room_id = r._id JOIN location l ON r.location_id = l._id JOIN user u ON res.user_id = u._id JOIN time_slot t ON res.time_slot_id = t._id JOIN recurrence rec on res.recurrence_id = rec._id JOIN recurrence_type rect on rec.recurrence_type_id = rect._id WHERE date >= CURDATE()" . " ORDER BY date";
         }
         echo "</h3>" . "\r\n";
 
@@ -54,7 +54,12 @@ class ViewMeetings {
      	        echo '         <td>' . $row['start_time'] . '</td>' . "\r\n";
                 echo '         <td>' . $row['end_time'] . '</td>' . "\r\n";
                 echo '         <td>' . $row['description'] . '</td>' . "\r\n";
-                echo '         <td>' . "<a class='home_page_link' onclick='return confirm(\"Are you sure?\")'href='cancel_reservation.php?reservation_id=" . $row['_id']  . "'>Delete</a>" . '</td>' . "\r\n";
+                if ($row['recurrance_id'] != 1) {
+                    echo '<td>' . "<a class='home_page_link' onclick='return confirm(\"Are you sure?\")'href='cancel_reservation.php?reservation_id=" . $row['_id']  . "'>Delete</a>" . ' [' ."<a class='home_page_link' onclick='return confirm(\"This will remove all reservations in the recurrance. Are you sure this is what you want?\")'href='cancel_reservation.php?recurrance_id=" . $row['recurrance_id']  . "'>All</a>". ']</td>' . "\r\n";
+                } else {
+                    echo '<td>' . "<a class='home_page_link' onclick='return confirm(\"Are you sure?\")'href='cancel_reservation.php?reservation_id=" . $row['_id']  . "'>Delete</a>" . '</td>' . "\r\n";
+
+                }
      	        echo '      </tr>' . "\r\n";
             }
